virtual opengl matrix stack, light glare billboards

+++ stars.cpp
@@ -18,7 +18,7 @@ cStars::render() {
     ::glPushMatrix();
 
     ::glRotatef( m_latitude, 1.0f, 0.0f, 0.0f ); // ustawienie osi OY na północ
-    ::glRotatef( -std::fmod( Global::fTimeAngleDeg, 360.0f ), 0.0f, 1.0f, 0.0f ); // obrót dobowy osi OX
+    ::glRotatef( -std::fmod( (float)Global::fTimeAngleDeg, 360.0f ), 0.0f, 1.0f, 0.0f ); // obrót dobowy osi OX
 
     ::glPointSize( 2.0f );
     if( Global::bUseVBO ) {
